adns: renamed update() to run()
Moved response delivery loop out of process_fd() and into update(),
which, by the way, was renamed to run(). Arguably, delivering the
events in process_fd() is fine, but then the loop would need to
be run in occasion of timeouts as well. The way it is now, events
are delivered in run() and the i/o callbacks merely update ADNS'
internal state.