test-dns: schedule needs to run twice.
The call to dns::run() might register a new timeout, so schedule::run()
has to be called afterwards to figure out the correct number of seconds
to wait. However, schedule::run() also delivers those timeouts, so it
has to run before dns::run() is called. This is unsatisfactory.
Arguably, it doesn't matter because schedule::run() is a very cheap
call, it's just that the structure looks odd.