build: configured installing and distribution of doxygen docs
A funny thing about Automake is that it needs an exact list of files
that ought to be distributed and installed. This list is, of course, not
exactly ease to provide if those files are generated with an external
tool like Doxygen. In the end, it turned out that disabling Doxygen in
the 'distcheck' target (which happens to be the offender) is the best
way to remedy the problem:

  http://permalink.gmane.org/gmane.comp.sysutils.automake.general/9417